feat(can_asc): basic support of CAN files in asc format
Support opening of CAN files in ".asc" format for both
remote and convert command.
Messages are converted to DLT as NwTrace/CAN msgs. The ECU id is
CAN<x> (x as CAN bus/channel id).
Apid and ctid are fixed to CAN/TC.
A provided BusMapping inside the .asc file is used as apid description.