feat(suspend): add detection of suspend/resume lifecycles
Added a detection for suspend/resume lifecycles from ECUs using suspend-to-ram.
Those lifecycles start after some time gap with the timestamp monotonically increasing
instead of starting from 0.
Currently the reception times need to be interrupted for at least 10s to be
identified / checked for a suspend/resume lifecycle.
The detection is important as otherwise the calculated absolute time for
those messages is wrong as the messages seem to belong to the prev. lifecycle
and thus the monotonic timestamp is added as a offset to the first lifecycle start time.