fix: resume lifecycles sometimes wrongly sorted
Resumed lifecycles might have a start_time earlier than the lifecycle the
resume was detected from.
This is a bit unexpected but not part of this PR.
In those cases the lifecycle was wrongly sorted.
For cmd remote the start_time is adjusted to 1us after the resumed one.
For cmd convert only the sort order is changed.