feat(remote): one_pass streams
Add first support to one pass streams.
The intended use-case is if e.g. a set of queries should be performed on a huge amount of input messages.
Currently the full messages would need to be loaded into RAM.
With the one_pass_stream feature the queries will be executed during the first read of the messages. Afterwards the messages are not kept in RAM.
This reduces the memory need.
fba-cli will use this to perform fishbone analysis on huge input files.