fix(filetransfer): speedup filetransfer state generation
The state generation was always generating the full tree
of transfer items incl. the sorted by name.

This had a O(n^2) complexity and was slowing down significantly
if lots of file transfers have been included in the logs.

Refactored the code to be called by each change and updating.
This reduces the times e.g. for a 6000 filetransfers example from 200s
to around 1s.

Additionally the "sort by name" was changed to "sort by ECU/name" with a
subtree for each ECU.
Sizes are now "pretty printed" using size crate.
Memory is still kept in RAM but tried to allocate on first data package
and via try_reserve_exact incl. error handling if it cannot be allocated.
(Still it will fill up the heap significantly if misused.)